+++ packages/base-controller/src/RestrictedMessenger.ts
@@ -316,7 +316,9 @@ export class RestrictedMessenger<
     SelectorReturnValue,
   >(
     event: EventType,
-    handler: ExtractEventHandler<Event, EventType>,
+    handler:
+      | ExtractEventHandler<Event, EventType>
+      | SelectorEventHandler<SelectorReturnValue>,
     selector?: SelectorFunction<Event, EventType, SelectorReturnValue>,
   ) {
     if (!this.#isAllowedEvent(event)) {
@@ -326,7 +328,10 @@ export class RestrictedMessenger<
     if (selector) {
       return this.#messenger.subscribe(event, handler, selector);
     }
-    return this.#messenger.subscribe(event, handler);
+    return this.#messenger.subscribe(
+      event,
+      handler as ExtractEventHandler<Event, EventType>,
+    );
   }
 
   /**
@@ -340,12 +345,19 @@ export class RestrictedMessenger<
    * @param handler - The event handler to unregister.
    * @throws Will throw if the given event is not an allowed event for this messenger.
    * @template EventType - A type union of allowed Event type strings.
+   * @template SelectorReturnValue - The selector return value.
    */
   unsubscribe<
     EventType extends
       | AllowedEvent
       | (Event['type'] & NamespacedName<Namespace>),
-  >(event: EventType, handler: ExtractEventHandler<Event, EventType>) {
+    SelectorReturnValue = unknown,
+  >(
+    event: EventType,
+    handler:
+      | ExtractEventHandler<Event, EventType>
+      | SelectorEventHandler<SelectorReturnValue>,
+  ) {
     if (!this.#isAllowedEvent(event)) {
       throw new Error(`Event missing from allow list: ${event}`);
     }